# Optimize CubicBez::arclen