Before
```
test bench_cubic_arclen_1e_4            ... bench:         281 ns/iter (+/- 28)
test bench_cubic_arclen_1e_5            ... bench:         281 ns/iter (+/- 20)
test bench_cubic_arclen_1e_6            ... bench:         608 ns/iter (+/- 26)
```

After
```
test bench_cubic_arclen_1e_4 ... bench:         267 ns/iter (+/- 13)
test bench_cubic_arclen_1e_5 ... bench:         256 ns/iter (+/- 12)
test bench_cubic_arclen_1e_6 ... bench:         562 ns/iter (+/- 56)
```

Looks like a statistically significant improvement.